pyproject.toml
[tool.poetry]
name = "l2-transfer-matrix-method"
version = "0.1.0"
description = ""
authors = ["Charlie Kennedy <charleslloydkennedy@gmail.com>"]
packages = [
    { include="transfer_matrix_method" }
]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
